+++ tests/component/test_odt_project_and_sbd_generation.py
@@ -163,3 +163,41 @@ def test_frequency_sweep_sbd_generated_from_project_obs_block(authrequests):
     sbd_id = project.obs_blocks[0].sbd_ids[-1]
     get_response = authrequests.get(f"{ODT_URL}/sbds/{sbd_id}")
     assert get_response.status_code == HTTPStatus.OK
+
+
+def test_survey_sbds_generated_from_project_obs_block(authrequests):
+    # First need to add a Project to generate from
+    project = TestDataFactory.project_with_two_low_targets(prj_id=None, prsl_ref=None)
+    post_response = authrequests.post(
+        f"{ODT_URL}/prjs",
+        data=project.model_dump_json(),
+        headers={"Content-Type": "application/json"},
+    )
+    assert post_response.status_code == HTTPStatus.OK, post_response.text
+    prj_id = post_response.json()["prj_id"]
+    obs_block_id = post_response.json()["obs_blocks"][0]["obs_block_id"]
+
+    # Generate survey SBDefinitions (limit to 6 rows for a quick test)
+    survey_input = {
+        "pointings_file_uri": "hex_relax_beams.sweet_as.csv",
+        "centre_frequency_mhz": 155.47,
+        "scan_duration_min": 5.0,
+        "num_subarray_beams": 2,
+        "num_scans": 3,
+        "max_rows": 6,
+    }
+    generate_response = authrequests.post(
+        f"{ODT_URL}/prjs/{prj_id}/{obs_block_id}/generateSurveySBDefinitions",
+        json=survey_input,
+    )
+
+    assert generate_response.status_code == HTTPStatus.OK, generate_response.text
+    project = Project.model_validate_json(generate_response.text)
+
+    # 6 targets with 2 beams * 3 scans = 6 targets per SBD => 1 SBD
+    assert len(project.obs_blocks[0].sbd_ids) == 1
+
+    # Check the SBDefinition exists in the ODA
+    sbd_id = project.obs_blocks[0].sbd_ids[0]
+    get_response = authrequests.get(f"{ODT_URL}/sbds/{sbd_id}")
+    assert get_response.status_code == HTTPStatus.OK